Added blacklist response setting

+++ app/src/main/java/com/tht/k3pler/frag/SettingsPageInflater.java
@@ -8,17 +8,16 @@ import android.content.res.Resources;
 import android.text.InputType;
 import android.text.SpannableString;
 import android.text.style.UnderlineSpan;
+import android.util.Log;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.WindowManager;
 import android.widget.EditText;
 import android.widget.LinearLayout;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.tht.k3pler.R;
 import com.tht.k3pler.handler.SqliteDBHelper;
-import com.tht.k3pler.sub.SQLiteBL;
 import com.tht.k3pler.sub.SQLiteSettings;
 
 import java.util.ArrayList;
@@ -30,8 +29,9 @@ public class SettingsPageInflater {
     private ViewGroup viewGroup;
     private SqliteDBHelper sqliteDBHelper;
     private String defaultBuffer = "10485760", defaultBufferNum = "10x1024x1024";
+    private String[] respOptions, matchOptions;
     // ** //
-    private TextView txvStPort, txvStMaxBuffer, txvStMatchType, txvStRestart;
+    private TextView txvStPort, txvStMaxBuffer, txvStMatchType, txvStBlResponse, txvStRestart;
 
     public SettingsPageInflater(Context context, ViewGroup viewGroup){
         this.context = context;
@@ -41,23 +41,32 @@ public class SettingsPageInflater {
         txvStPort = viewGroup.findViewById(R.id.txvStPort);
         txvStMaxBuffer = viewGroup.findViewById(R.id.txvStMaxBuffer);
         txvStMatchType = viewGroup.findViewById(R.id.txvStMatchType);
+        txvStBlResponse = viewGroup.findViewById(R.id.txvStBlResponse);
         txvStRestart = viewGroup.findViewById(R.id.txvStRestart);
+        respOptions = new String[]{
+                context.getString(R.string.resp_status_1),
+                context.getString(R.string.resp_status_2),
+                context.getString(R.string.resp_status_3),
+                context.getString(R.string.resp_status_4),};
+        matchOptions = new String[]{
+                context.getString(R.string.match_type_1),
+                context.getString(R.string.match_type_2)};
         setValues();
     }
     private void setValues(){
         ArrayList<String> settings = getSettings();
-        settextWithUnderline(txvStPort, settings.get(0));
+        setTextWithUnderline(txvStPort, settings.get(0));
         if(settings.get(1).equals(defaultBuffer)){
-            settextWithUnderline(txvStMaxBuffer, defaultBufferNum);
+            setTextWithUnderline(txvStMaxBuffer, defaultBufferNum);
         }else {
-            settextWithUnderline(txvStMaxBuffer, settings.get(1));
+            setTextWithUnderline(txvStMaxBuffer, settings.get(1));
         }
         if(Integer.valueOf(settings.get(2))==0){
-            settextWithUnderline(txvStMatchType, context.getString(R.string.match_type_1));
+            setTextWithUnderline(txvStMatchType, context.getString(R.string.match_type_1));
         }else{
-            settextWithUnderline(txvStMatchType, context.getString(R.string.match_type_2));
+            setTextWithUnderline(txvStMatchType, context.getString(R.string.match_type_2));
         }
-
+        setTextWithUnderline(txvStBlResponse, getRespText(Integer.valueOf(settings.get(3))));
         txvStPort.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -76,19 +85,28 @@ public class SettingsPageInflater {
             @Override
             public void onClick(View view) {
                 showSelectDialog(context.getString(R.string.lyt_match_type).replace(":", ""),
-                        txvStMatchType.getText().toString(), SQLiteSettings.MATCH_DATA);
+                        txvStMatchType.getText().toString(), SQLiteSettings.MATCH_DATA, matchOptions);
+            }
+        });
+        txvStBlResponse.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                showSelectDialog(context.getString(R.string.lyt_blacklist_response).replace(":", ""),
+                        txvStBlResponse.getText().toString(), SQLiteSettings.RESPONSE_DATA, respOptions);
             }
         });
     }
-    private void settextWithUnderline(TextView textView, String text){
+    private void setTextWithUnderline(TextView textView, String text){
         SpannableString content = new SpannableString(text);
         content.setSpan(new UnderlineSpan(), 0, content.length(), 0);
         textView.setText(content);
     }
     public ArrayList<String> getSettings(){
+        Boolean notFound = false;
         int port = Integer.parseInt(context.getString(R.string.default_port));
         int buffer = 10 * 1024 * 1024;
         int match =  Integer.parseInt(context.getString(R.string.default_matchtype));
+        int resp = 0;
         sqliteDBHelper = new SqliteDBHelper(context,
                 new SQLiteSettings(context).getWritableDatabase(),
                 SQLiteSettings.PORT_DATA, SQLiteSettings.TABLE_NAME);
@@ -96,6 +114,7 @@ public class SettingsPageInflater {
             port = Integer.parseInt(sqliteDBHelper.get(SQLiteSettings.PORT_DATA));
         }catch (NumberFormatException e){
             e.printStackTrace();
+            notFound = true;
         }
         sqliteDBHelper = new SqliteDBHelper(context,
                 new SQLiteSettings(context).getWritableDatabase(),
@@ -104,6 +123,7 @@ public class SettingsPageInflater {
             buffer = Integer.parseInt(sqliteDBHelper.get(SQLiteSettings.BUFFER_DATA));
         }catch (NumberFormatException e){
             e.printStackTrace();
+            notFound = true;
         }
         sqliteDBHelper = new SqliteDBHelper(context,
                 new SQLiteSettings(context).getWritableDatabase(),
@@ -112,18 +132,32 @@ public class SettingsPageInflater {
             match = Integer.parseInt(sqliteDBHelper.get(SQLiteSettings.MATCH_DATA));
         }catch (NumberFormatException e){
             e.printStackTrace();
+            notFound = true;
         }
-        sqliteDBHelper.deleteAll();
-        HashMap<String, String> hashMap = new HashMap<>();
-        hashMap.put(SQLiteSettings.PORT_DATA, String.valueOf(port));
-        hashMap.put(SQLiteSettings.BUFFER_DATA, String.valueOf(buffer));
-        hashMap.put(SQLiteSettings.MATCH_DATA, String.valueOf(match));
-        sqliteDBHelper.insertMultiple(hashMap);
+        sqliteDBHelper = new SqliteDBHelper(context,
+                new SQLiteSettings(context).getWritableDatabase(),
+                SQLiteSettings.RESPONSE_DATA, SQLiteSettings.TABLE_NAME);
+        try {
+            resp = Integer.parseInt(sqliteDBHelper.get(SQLiteSettings.RESPONSE_DATA));
+        }catch (NumberFormatException e){
+            e.printStackTrace();
+            notFound = true;
+        }
+        if(notFound){
+            sqliteDBHelper.deleteAll();
+            HashMap<String, String> hashMap = new HashMap<>();
+            hashMap.put(SQLiteSettings.PORT_DATA, String.valueOf(port));
+            hashMap.put(SQLiteSettings.BUFFER_DATA, String.valueOf(buffer));
+            hashMap.put(SQLiteSettings.MATCH_DATA, String.valueOf(match));
+            hashMap.put(SQLiteSettings.RESPONSE_DATA, String.valueOf(resp));
+            sqliteDBHelper.insertMultiple(hashMap);
+        }
         sqliteDBHelper.close();
         ArrayList<String> arrayList = new ArrayList<>();
         arrayList.add(String.valueOf(port));
         arrayList.add(String.valueOf(buffer));
         arrayList.add(String.valueOf(match));
+        arrayList.add(String.valueOf(resp));
         return arrayList;
     }
 
@@ -166,11 +200,12 @@ public class SettingsPageInflater {
         return (int) (dp * Resources.getSystem().getDisplayMetrics().density);
     }
     @SuppressWarnings("deprecation")
-    private void showSelectDialog(String title, String currentItem, final String DATA){
-        if(currentItem.equals(context.getString(R.string.match_type_1))){currentItem="0";}else{currentItem="1";}
+    private void showSelectDialog(String title, String currentItem, final String DATA, String[] options){
+        if(DATA.equals(SQLiteSettings.MATCH_DATA)){
+            if(currentItem.equals(context.getString(R.string.match_type_1))){currentItem="0";}else{currentItem="1";}
+        }
+        if(DATA.equals(SQLiteSettings.RESPONSE_DATA)){currentItem = String.valueOf(getRespID(currentItem));}
         final String currentItem1 = currentItem;
-        final String[] options = new String[]{context.getString(R.string.match_type_1),
-                context.getString(R.string.match_type_2)};
         AlertDialog.Builder builder = new AlertDialog.Builder(context, android.R.style.Theme_DeviceDefault_Dialog);
         builder.setTitle(title);
         builder.setItems(options, new DialogInterface.OnClickListener() {
@@ -188,4 +223,31 @@ public class SettingsPageInflater {
         alertDialog.getWindow().setType(WindowManager.LayoutParams.TYPE_SYSTEM_ALERT);
         alertDialog.show();
     }
+    private int getRespID(String resp){
+        if(resp.equals(context.getString(R.string.resp_status_1)))
+            return 0;
+        else if(resp.equals(context.getString(R.string.resp_status_2)))
+            return 1;
+        else if(resp.equals(context.getString(R.string.resp_status_3)))
+            return 2;
+        else if(resp.equals(context.getString(R.string.resp_status_4)))
+            return 3;
+        else
+            return -1;
+    }
+    private String getRespText(int ID){
+        switch (ID){
+            case 0:
+                return context.getString(R.string.resp_status_1);
+            case 1:
+                return context.getString(R.string.resp_status_2);
+            case 2:
+                return context.getString(R.string.resp_status_3);
+            case 3:
+                return context.getString(R.string.resp_status_4);
+            default:
+                return "0";
+        }
+    }
+
 }
